created new file and added functions
Separated listeners from the main function to change the theme. Added function to change background and change text. Adding more colors to the ls
