+++ index.html
@@ -25,11 +25,11 @@
   <body>
     <header>
       <nav>
-        <ul class="darkText">
-          <li><a href="#projects">Projects</a></li>
-          <li><a href="#about">About</a></li>
-          <li><a href="#contact">Contact</a></li>
-          <li><a href="palettes/">Colours</a></li>
+        <ul>
+          <li><a href="#projects" class="darkText">Projects</a></li>
+          <li><a href="#about" class="darkText">About</a></li>
+          <li><a href="#contact" class="darkText">Contact</a></li>
+          <li><a href="palettes/" class="darkText">Colours</a></li>
         </ul>
       </nav>
     </header>
@@ -61,13 +61,13 @@
               <a
                 href="https://ephemeral-sundae-92c10d.netlify.app/"
                 target="_blank"
-                class="demo-button button"
+                class="lightButton button"
                 >Demo</a
               >
               <a
                 href="https://github.com/Noroff-FEU-Assignments/cross-course-project-Spekkhogger"
                 target="_blank"
-                class="code-button button"
+                class="button darkButton"
                 >Code</a
               >
             </div>
@@ -93,13 +93,13 @@
               <a
                 href="https://wonderful-sorbet-adb43f.netlify.app/"
                 target="_blank"
-                class="demo-button button"
+                class="lightButton button"
                 >Demo</a
               >
               <a
                 href="https://github.com/Spekkhogger/ingeborg_semester-project"
                 target="_blank"
-                class="code-button button"
+                class="darkButton button"
                 >Code</a
               >
             </div>
@@ -121,13 +121,13 @@
               <a
                 href="https://resilient-creponne-254619.netlify.app/"
                 target="_blank"
-                class="demo-button button"
+                class="button lightButton"
                 >Demo</a
               >
               <a
                 href="https://github.com/Noroff-FEU-Assignments/project-exam-1-Spekkhogger"
                 target="_blank"
-                class="code-button button"
+                class="darkButton button"
                 >Code</a
               >
             </div>
@@ -136,7 +136,7 @@
       </div>
       <span id="about"></span>
 
-      <div class="about-me-container ligthText">
+      <div class="about-me-container ligthText boldColor">
         <h2>About me</h2>
         <div class="about-wrap">
           <div class="about-left">
@@ -160,10 +160,10 @@
             <img src="images/IMG_5683 Medium.jpeg" alt="picture of me" />
           </div>
         </div>
-        <a href="https://github.com/Spekkhogger" class="download button"
+        <a href="https://github.com/Spekkhogger" class="download button lightButton"
           >GitHub</a
         >
-        <a href="#" class="download button">Download resume (pdf)</a>
+        <a href="#" class="download button lightButton">Download resume (pdf)</a>
       </div>
 
       <div class="contact-container" id="contact">
