+++ palettes/index.html
@@ -192,6 +192,6 @@
     </main>
 </body>
 <script src="../copiText.js"></script>
-<script src="../mouseOverEvent.js"></script>
+<script type="module" src="../addPalette.mjs"></script>
 <script src="../changeTheme.mjs" type="module"></script>
 </html>
