+++ style.css
@@ -26,7 +26,6 @@ nav {
     top: 0;
     right: 0;
     min-width: 100%;
-    background-color: var(--background-color);
     z-index: 1;
 }
 
@@ -57,7 +56,6 @@ main {
     font-family: 'Comfortaa', cursive;
     font-size: 4rem;
     position: relative;
-    color: var(--dark-brown);
     margin: 0px 0px 0px;
     z-index: 999;
 }
@@ -118,7 +116,6 @@ main {
 
 .button {
     padding: 10px 35px;
-    color: var(--dark-brown);
     border-radius: 5px;
 }
 
@@ -128,26 +125,28 @@ main {
     gap: 1.5em;
 }
 
-.demo-button {
+.lightButton {
     background-color: var(--beige);
     border: 2px solid var(--beige);
+    color: var(--dark-blue);
 }
 
-.demo-button:hover {
-    background-color: var(--background-color);
-    transition: ease-in-out 0.2s; 
+.lightButton:hover {
+    text-decoration: line-through;
+    opacity: 90%;
+    transition: ease-in-out 0.2s;
 }
 
-.code-button {
+.darkButton {
     background-color: var(--dark-blue);
     border: 2px solid var(--dark-blue);
     color: var(--background-color);
 }
 
-.code-button:hover {
-    background-color: var(--background-color);
-    transition: ease-in-out 0.2s; 
-    color: var(--dark-blue);
+.darkButton:hover {
+    text-decoration: line-through;
+    opacity: 90%;
+    transition: ease-in-out 0.2s;
 }
 
 
@@ -159,10 +158,14 @@ main {
     display: flex;
     flex-direction: column;
     align-items: center;
-    background-color: var(--dark-blue);
+    
     color: var(--background-color);
     min-width: 100%;
     padding: 2em;
+}
+
+.boldColor {
+    background-color: var(--dark-blue);
 }
 
 .about-wrap {
@@ -191,20 +194,6 @@ main {
 #about-list li{
     margin: 10px;
 }
-
-.download {
-    background-color: var(--beige);
-    border: 2px solid var(--beige);
-    margin: 10px 0px;
-}
-
-.download:hover {
-    background-color: var(--dark-blue);
-    color: var(--beige);
-    transition: ease-in-out 0.2s;
-}
-
-
 
 /*Contact section*/
